Add new incident scenario for burning small aircraft
This commit introduces a thrilling new scenario for a burning small aircraft, because who doesn't want to deal with that? The new entry is added at index 335 with the necessary response units: 8 fire engines, 2 emergency vehicles, 1 rescue vehicle, and more. Just what we needed to spice up the incident list!