Add autoAcceptMaxDistance setting for mission filtering
This commit introduces a new setting, 'autoAcceptMaxDistance', which allows users to specify the maximum distance for missions to be automatically accepted. The default value is set to 9999. Additionally, the autoAccept function has been updated to filter missions based on this new setting, enhancing the script's functionality and user control.

Because who doesn't want to accept missions from a distance?