Add new scene for goal scored incident
This commit introduces a new scene for when a goal is scored in the game, identified by the code "375". The scene requires one fire truck (LF) to respond. This addition enhances the gameplay experience by providing more realistic scenarios for players to manage.

Because who doesn't love a good goal celebration?