Add new scene for wasp attack incident
This commit introduces a new scene (ID: 754) for a wasp attack incident, enhancing the variety of emergency scenarios available in the game. The scene includes the deployment of a rescue vehicle (RTW) and a medical emergency vehicle (NEF) to respond to the incident.

Now you can save the day from pesky wasps!