Add buyExtensions function for vehicle extension management
This commit introduces the buyExtensions function to facilitate the purchase of vehicle extensions in the Leitstellenspiel game. The function allows users to specify the extension ID, building type, and range of buildings to process. It constructs the necessary request headers and handles the asynchronous fetching of extension credits for the specified buildings. Additionally, helper functions for buying specific types of extensions (fire brigade, police, THW, SEG) are also defined.

Now you can extend your vehicles like a pro, because who wouldn't want to be the ultimate vehicle manager?