Add vehicle assignments for scene 430: Ammoniakaustritt in Eishalle
This commit introduces vehicle assignments for scene 430, which includes a significant increase in the number of vehicles required for an ammonia leak scenario. The new assignments are as follows: 15 LF, 3 ELW, 1 ELW2, 2 DEKONP, 2 GWM, 2 GWG, 1 GWA, 1 DLK, 2 SW, 4 RW, 8 POL, and 1 RTW.

Prepare for a chilling response to an ammonia leak in the ice hall!